#add check device script
