+++ mbti_personality_eda.py
@@ -5,7 +5,7 @@ from torch.optim import AdamW
 from torch.utils.data import DataLoader
 from transformers import AutoTokenizer, get_scheduler
 
-from models.deep_models.utils.tools import get_device
+from models.deep_models.utils.check_device import get_device
 from projects.mbti_personality.load_data import load_data
 from projects.mbti_personality.model import LanguageModel
